refactor(deprecate): extract _properties + messages
- move `_DeprecatedProperty` and `_StrictProperty` into new `_properties.py` (property descriptors, no call-forwarding-engine coupling)
- move warning templates, `_validate_template_mgs`, and the emitters (`_raise_warn*`, `_consume_warn_budget`, `_source_display_name`) into new `messages.py`, shared by `callables` and `proxy`
- relocate the generic `_unwrap_descriptor_target` to `utils.py`, the lowest common module for `messages` and the engine
- update the `deprecation.py` shim to re-export from the new homes; point `proxy`, `audit`, `__init__`, and the property unit test at `messages` / `_properties` directly

---
